requirements.txt
moviepy==1.0.3
numpy==1.22.3
Pillow==9.1.0
scikit-learn==1.0.2
tqdm==4.64.0

